+++ responses/article.go
@@ -33,7 +33,7 @@ type SingleArticleResponse struct {
 
 type ArticleListResponse struct {
 	Articles      []*ArticleResponse `json:"articles"`
-	ArticlesCount int                `json:"articlesCount"`
+	ArticlesCount int64              `json:"articlesCount"`
 }
 
 func NewArticleResponse(c echo.Context, a *model.Article) *SingleArticleResponse {
@@ -61,7 +61,7 @@ func NewArticleResponse(c echo.Context, a *model.Article) *SingleArticleResponse
 	return &SingleArticleResponse{ar}
 }
 
-func NewArticleListResponse(us *services.UserService, userID uint, articles []model.Article, count int) *ArticleListResponse {
+func NewArticleListResponse(us *services.UserService, userID uint, articles []model.Article, count int64) *ArticleListResponse {
 	r := new(ArticleListResponse)
 	r.Articles = make([]*ArticleResponse, 0)
 	for _, a := range articles {
